+++ CHANGELOG.md
@@ -37,6 +37,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Force plugins to be added only once (#422, **@RiscadoA**).
 - Return direct references to resources instead of guards on World (#922, **@RiscadoA**).
 - Stop using old serialization to load Settings (#1100, **@RiscadoA**).
+- Make transform gizmo screen size indpendent from entity's distance to camera (#1017, **@DiogoMendonc-a**)
 
 ### Changed